Fix entity validation errors
