+++ jest.env.js
@@ -1,1 +1,2 @@
 process.env.RUNNING_TESTS = 'true';
+process.env.ENABLE_GRAPH_OBJECT_SCHEMA_VALIDATION = '1';
